Replace React Router v4 with Redux-First Router
This commit has the nice side effect of having Hot Module Reloading
actually working. For some reason before the entire page would refresh,
resulting in state being wiped and having to fetch the data again.

Now it just works. This is a bigger commit since it was done in its own
branch. I had some problems though so the squashed commit does not look
like a proper merge.

This is also the last commit that will rewrite history in the future.

Signed-off-by: Lauritz Hilsøe <mail@lauritz.me>